Move AP to its own mini package
