+++ src/ActivityPub/Api/SiteInbox.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Smolblog\Api\ActivityPub;
+namespace Smolblog\ActivityPub\Api;
 
 use Smolblog\Api\Endpoint;
 use Smolblog\Api\EndpointConfig;
