+++ src/ActivityPub/Api/Webfinger.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Smolblog\Api\ActivityPub;
+namespace Smolblog\ActivityPub\Api;
 
 use Smolblog\Api\ApiEnvironment;
 use Smolblog\Api\Endpoint;
